+++ maasaisearchproxy-20260424/alibabacloud_maasaisearchproxy20260424/models/__init__.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 from ._web_search_request import WebSearchRequest
 from ._web_search_response_body import WebSearchResponseBody
 from ._web_search_response import WebSearchResponse
+from ._web_search_response_body import WebSearchResponseBodyDataResultSource
 from ._web_search_response_body import WebSearchResponseBodyDataResult
 from ._web_search_response_body import WebSearchResponseBodyData
 
@@ -12,6 +13,7 @@ __all__ = [
     WebSearchRequest,
     WebSearchResponseBody,
     WebSearchResponse,
+    WebSearchResponseBodyDataResultSource,
     WebSearchResponseBodyDataResult,
     WebSearchResponseBodyData
 ]