+++ maasaisearchproxy-20260424/alibabacloud_maasaisearchproxy20260424/models/_web_search_response_body.py
@@ -13,13 +13,10 @@ class WebSearchResponseBody(DaraModel):
         code: int = None,
         data: main_models.WebSearchResponseBodyData = None,
         message: str = None,
-        trace_id: str = None,
     ):
         self.code = code
         self.data = data
         self.message = message
-        # requestId
-        self.trace_id = trace_id
 
     def validate(self):
         if self.data:
@@ -39,9 +36,6 @@ class WebSearchResponseBody(DaraModel):
         if self.message is not None:
             result['message'] = self.message
 
-        if self.trace_id is not None:
-            result['traceId'] = self.trace_id
-
         return result
 
     def from_map(self, m: dict = None):
@@ -55,9 +49,6 @@ class WebSearchResponseBody(DaraModel):
 
         if m.get('message') is not None:
             self.message = m.get('message')
-
-        if m.get('traceId') is not None:
-            self.trace_id = m.get('traceId')
 
         return self
 
@@ -104,29 +95,38 @@ class WebSearchResponseBodyData(DaraModel):
 
         return self
 
-
-
 class WebSearchResponseBodyDataResult(DaraModel):
     def __init__(
         self,
+        date: str = None,
         snippet: str = None,
+        source: main_models.WebSearchResponseBodyDataResultSource = None,
         title: str = None,
         url: str = None,
     ):
+        self.date = date
         self.snippet = snippet
+        self.source = source
         self.title = title
         self.url = url
 
     def validate(self):
-        pass
+        if self.source:
+            self.source.validate()
 
     def to_map(self):
         result = dict()
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.date is not None:
+            result['date'] = self.date
+
         if self.snippet is not None:
             result['snippet'] = self.snippet
+
+        if self.source is not None:
+            result['source'] = self.source.to_map()
 
         if self.title is not None:
             result['title'] = self.title
@@ -138,14 +138,66 @@ class WebSearchResponseBodyDataResult(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('date') is not None:
+            self.date = m.get('date')
+
         if m.get('snippet') is not None:
             self.snippet = m.get('snippet')
+
+        if m.get('source') is not None:
+            temp_model = main_models.WebSearchResponseBodyDataResultSource()
+            self.source = temp_model.from_map(m.get('source'))
 
         if m.get('title') is not None:
             self.title = m.get('title')
 
         if m.get('url') is not None:
             self.url = m.get('url')
+
+        return self
+
+
+
+class WebSearchResponseBodyDataResultSource(DaraModel):
+    def __init__(
+        self,
+        domain: str = None,
+        favicon: str = None,
+        name: str = None,
+    ):
+        self.domain = domain
+        self.favicon = favicon
+        self.name = name
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.domain is not None:
+            result['domain'] = self.domain
+
+        if self.favicon is not None:
+            result['favicon'] = self.favicon
+
+        if self.name is not None:
+            result['name'] = self.name
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('domain') is not None:
+            self.domain = m.get('domain')
+
+        if m.get('favicon') is not None:
+            self.favicon = m.get('favicon')
+
+        if m.get('name') is not None:
+            self.name = m.get('name')
 
         return self