+++ maasaisearchproxy-20260424/ChangeLog.md
@@ -1,3 +1,9 @@
+2026-05-14 Version: 2.0.0
+- Update API WebSearch: add response parameters Body.data.result.$.date.
+- Update API WebSearch: add response parameters Body.data.result.$.source.
+- Update API WebSearch: delete response parameters Body.traceId.
+
+
 2026-05-12 Version: 1.0.0
 - Generated python 2026-04-24 for MaasAISearchProxy.